the specification has been completed and bugs have been fixed

+++ src/main/java/ru/tyutterin/coffeemaker/model/entity/CoffeeResidue.java
@@ -22,9 +22,9 @@ public class CoffeeResidue {
     private CoffeeMaker coffeeMaker;
     @CreationTimestamp
     private LocalDateTime fallTime;
-    private int quantity;
+    private long quantity;
 
-    public CoffeeResidue(CoffeeMaker coffeeMaker, int quantity) {
+    public CoffeeResidue(CoffeeMaker coffeeMaker, long quantity) {
         this.coffeeMaker = coffeeMaker;
         this.quantity = quantity;
     }
