+++ src/main/java/ru/tyutterin/coffeemaker/model/entity/WaterResidue.java
@@ -22,9 +22,9 @@ public class WaterResidue {
     private CoffeeMaker coffeeMaker;
     @CreationTimestamp
     private LocalDateTime fallTime;
-    private int quantity;
+    private long quantity;
 
-    public WaterResidue(CoffeeMaker coffeeMaker, int quantity) {
+    public WaterResidue(CoffeeMaker coffeeMaker, long quantity) {
         this.coffeeMaker = coffeeMaker;
         this.quantity = quantity;
     }
